requirements.txt
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
python-dotenv==1.0.0
tavily-python==0.2.6
requests==2.31.0